new-york-collisions/requirements.txt
geopandas
missingno
numpy
pandas
shapelystreamlit
